+++ vite.config.js
@@ -1,0 +1,19 @@
+import { defineConfig } from 'vite'
+
+export default defineConfig({
+  publicDir: 'public',
+  build: {
+    outDir: 'dist',
+    assetsDir: 'assets',
+    rollupOptions: {
+      input: {
+        main: './index.html'
+      }
+    }
+  },
+  server: {
+    port: 5173,
+    open: true
+  }
+})
+